Avoid casting to float, fixed Math.Pow

+++ Runtime/Model/Session/SessionSubmissionJob.cs
@@ -2,7 +2,7 @@
 {
     internal sealed class SessionSubmissionJob
     {
-        public float NextInvokeTime { get; set; }
+        public double NextInvokeTime { get; set; }
         public UniqueEvent[] UniqueEvents { get; set; }
         public SessionEvent[] SessionEvents { get; set; }
         public uint NumberOfRetries { get; set; }
